Starting rework of ceptic-python, moved/renamed some old files, implemented net and encode modules, started work on common.py, created unit tests for encode module
