+++ ceptic/stream.py
@@ -1,0 +1,4 @@
+
+
+class StreamHandler(object):
+    pass
