+++ tests/fixtures/client.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ceptic.client import CepticClient, client_settings
+from oldceptic.client import CepticClient, client_settings
 
 
 @pytest.fixture(scope="function")
