+++ tests/oldintegration_command_test.py
@@ -7,9 +7,9 @@ if version_info < (3, 0):  # if running python 2
 else:
     from queue import Queue
 
-from ceptic.server import server_settings, begin_exchange
-from ceptic.client import client_settings
-from ceptic.common import CepticResponse, CepticException, CepticStatusCode
+from oldceptic.server import server_settings, begin_exchange
+from oldceptic.client import client_settings
+from oldceptic.common import CepticResponse, CepticException, CepticStatusCode
 
 
 # TESTS:
